--- crawling/DB_FRUITS.csv
@@ -1,1 +1,0 @@
-﻿
